deps: upgrade reqwest 0.13, rusqlite 0.39, alloy 1.7, actix-governor 0.10, rand 0.10, getrandom 0.4
Addresses crates.io dependency audit warnings:
- reqwest 0.12 -> 0.13
- rusqlite 0.32 -> 0.39 (fix u64 FromSql removal: use i64 + cast)
- alloy 1 -> 1.7 (precise version)
- actix-governor 0.8 -> 0.10
- actix-web 4 -> 4.13 (precise version)
- rand 0.9 -> 0.10
- getrandom 0.3 -> 0.4
- Rust toolchain updated to 1.94

+++ crates/tempo-x402-soul/src/db.rs
@@ -736,13 +736,13 @@ impl SoulDatabase {
             params![fingerprint, now],
         )?;
 
-        let count: u64 = conn.query_row(
+        let count: i64 = conn.query_row(
             "SELECT count FROM pattern_counts WHERE fingerprint = ?1",
             params![fingerprint],
             |row| row.get(0),
         )?;
 
-        Ok(count)
+        Ok(count as u64)
     }
 
     /// Get pattern counts for multiple fingerprints.
@@ -780,8 +780,8 @@ impl SoulDatabase {
         let mut result = HashMap::new();
         let rows = stmt.query_map(params_refs.as_slice(), |row| {
             let fp: String = row.get(0)?;
-            let count: u64 = row.get(1)?;
-            Ok((fp, count))
+            let count: i64 = row.get(1)?;
+            Ok((fp, count as u64))
         })?;
         for row in rows {
             let (fp, count) = row?;
@@ -1292,12 +1292,12 @@ impl SoulDatabase {
                 "lock poisoned".into(),
             ))
         })?;
-        let count: u64 = conn.query_row(
+        let count: i64 = conn.query_row(
             "SELECT COUNT(*) FROM plans WHERE status = ?1",
             params![status],
             |row| row.get(0),
         )?;
-        Ok(count)
+        Ok(count as u64)
     }
 
     /// Force a WAL checkpoint to reclaim disk space.
@@ -1695,8 +1695,8 @@ impl SoulDatabase {
                 "lock poisoned".into(),
             ))
         })?;
-        let count: u64 = conn.query_row("SELECT COUNT(*) FROM goals", [], |row| row.get(0))?;
-        Ok(count)
+        let count: i64 = conn.query_row("SELECT COUNT(*) FROM goals", [], |row| row.get(0))?;
+        Ok(count as u64)
     }
 
     /// Get recently abandoned/failed goals (for retread detection).
@@ -2102,12 +2102,12 @@ impl SoulDatabase {
                 "lock poisoned".into(),
             ))
         })?;
-        let count: u64 = conn.query_row(
+        let count: i64 = conn.query_row(
             "SELECT COUNT(*) FROM plan_outcomes WHERE status = ?1",
             params![status],
             |row| row.get(0),
         )?;
-        Ok(count)
+        Ok(count as u64)
     }
 
     // ── Capability event operations ──
